+++ src/components/Login/Login.test.jsx
@@ -1,4 +1,4 @@
-import { render } from "@testing-library/react";
+import { render, prettyDOM, getByText } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { BrowserRouter as Router } from "react-router-dom";
 import LoginComponent from "./Login";
@@ -23,32 +23,52 @@ beforeEach(() => {
   );
 });
 
+test("Enter Username & Password are displayed correctly", () => {
+  const enterUsernameElement = renderScreen.getByText(/enter username/i);
+  expect(enterUsernameElement.textContent).toBe("Enter Username");
+
+  const enterPasswordElement = renderScreen.getByText(/Enter Password/i);
+  expect(enterPasswordElement.textContent).toBe("Enter Password");
+});
+
+test("Inputs should be blank on initial render", () => {
+  const usernameInputElement = renderScreen.getByRole("textbox", {
+    name: /enter username/i,
+  });
+  const passwordInputElement = renderScreen.getByLabelText(/password/i);
+
+  expect(usernameInputElement.value).toBe("");
+  expect(passwordInputElement.value).toBe("");
+});
+
+test("Username and Password are entered correctly", () => {
+  const usernameInputElement = renderScreen.getByRole("textbox", {
+    name: /enter username/i,
+  });
+  const passwordInputElement = renderScreen.getByLabelText(/password/i);
+
+  userEvent.type(usernameInputElement, "dlubbers");
+  userEvent.type(passwordInputElement, "passtest");
+
+  expect(usernameInputElement.value).toBe("dlubbers");
+  expect(passwordInputElement.value).toBe("passtest");
+});
+
 test("Error messages show on screen when username or password is incorrect", () => {
-  mockSubmit.mockImplementation((e) => e.preventDefault());
+  // mockSubmit.mockImplementation((e) => e.preventDefault());
 
-  const error = {
-    usernameError: "The username you have submitted is incorrect!",
-    passwordError: "The password you have submitted is incorrect!",
-  };
-
-  const enterUsernameElement = renderScreen.getByRole(
-    "textbox",
-    /Enter Username/i
-  );
-
-  const usernameErrorElement = renderScreen.queryByText(
+  const usernameInputElement = renderScreen.getByRole("textbox", {
+    name: /enter username/i,
+  });
+  const userErrorMessageElement = renderScreen.queryByText(
     /The username you have submitted is incorrect!/i
   );
+  const submitBtn = renderScreen.getByRole("button", { type: /submit/i });
 
-  // expect(usernameErrorElement).not.toBeInTheDocument();
+  expect(userErrorMessageElement).not.toBeInTheDocument();
 
-  const submitBtn = renderScreen.getByRole("button", { type: "submit" });
-
-  // userEvent.type(enterUsernameElement, "wrong username");
-
+  userEvent.type(usernameInputElement, "dlubbers");
   userEvent.click(submitBtn);
 
-  expect(mockSubmit).toBeCalledTimes(1);
-  // expect(enterUsernameElement.value).not.toBe("dlubbers");
-  // expect(usernameErrorElement.textContent).toBe(error.usernameError);
+  expect(userErrorMessageElement).toBeInTheDocument();
 });
