+++ src/pages/login.test.jsx
@@ -1,9 +1,11 @@
-import { render } from "@testing-library/react";
+import { prettyDOM, render, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { BrowserRouter as Router } from "react-router-dom";
 import Login from "./login";
 
 let renderScreen;
+
+const mockSubmit = jest.fn();
 
 beforeEach(() => {
   renderScreen = render(
@@ -21,25 +23,50 @@ test("Enter Username & Password are displayed correctly", () => {
   expect(enterPasswordElement.textContent).toBe("Enter Password");
 });
 
-test("Inputs should be blank on initial render", () => {
-  const usernameInputElement = renderScreen.getByRole("textbox");
-  expect(usernameInputElement.value).toBe("");
-  const passwordInputElement = renderScreen.getByLabelText(/password/i);
-  expect(passwordInputElement.value).toBe("");
-});
+// test("Inputs should be blank on initial render", () => {
+//   const usernameInputElement = renderScreen.getByRole("textbox", {
+//     name: /enter username/i,
+//   });
+//   expect(usernameInputElement.value).toBe("");
+//   const passwordInputElement = renderScreen.getByLabelText(/password/i);
+//   expect(passwordInputElement.value).toBe("");
+// });
 
-test("Username and Password are entered correctly", () => {
-  const username = "dlubbers";
-  const password = "passtest";
+// test("Username and Password are entered correctly", () => {
+//   const username = "dlubbers";
+//   const password = "passtest";
 
-  const usernameInputElement = renderScreen.getByRole("textbox", {
-    name: /username/i,
-  });
-  const passwordInputElement = renderScreen.getByLabelText("Enter Password");
+//   const usernameInputElement = renderScreen.getByRole("textbox", {
+//     name: /enter username/i,
+//   });
+//   const passwordInputElement = renderScreen.getByLabelText(/password/i);
 
-  userEvent.type(usernameInputElement, "dlubbers");
-  userEvent.type(passwordInputElement, "passtest");
+//   userEvent.type(usernameInputElement, "dlubbers");
+//   userEvent.type(passwordInputElement, "passtest");
 
-  expect(usernameInputElement.value).toBe(username);
-  expect(passwordInputElement.value).toBe(password);
-});
+//   expect(usernameInputElement.value).toBe(username);
+//   expect(passwordInputElement.value).toBe(password);
+// });
+
+// test("Error messages", () => {
+//   mockSubmit.mockImplementation((e) => e.preventDefault());
+//   console.log(prettyDOM(renderScreen.container));
+
+//   const usernameInputElement = renderScreen.getByRole(
+//     "textbox",
+//     /enter username/i
+//   );
+//   const passwordInputElement = renderScreen.getByLabelText(/password/i);
+
+//   const submitBtn = renderScreen.getByRole("button", {
+//     type: "submit",
+//   });
+
+//   console.log(submitBtn);
+
+//   userEvent.type(usernameInputElement, "dlubbers");
+//   userEvent.type(passwordInputElement, "test");
+//   userEvent.click(submitBtn);
+//   expect(mockSubmit).toBeCalledTimes(1);
+//   // expect(enterUsernameElement.value).not.toBe("dlubbers");
+// });
